% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{epiworldRenv}
\alias{epiworldRenv}
\title{Access to the epiworldR environment
This function is for internal use only.}
\usage{
epiworldRenv()
}
\value{
Returns the epiworldR environment
}
\description{
Access to the epiworldR environment
This function is for internal use only.
}
