% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{epiworldRShiny}
\alias{epiworldRShiny}
\alias{run_app}
\title{epiworldRShiny App Launcher}
\usage{
epiworldRShiny(...)

run_app(...)
}
\arguments{
\item{...}{Currently ignored.}
}
\value{
Loads and opens the RShiny app for the epiworldR package
}
\description{
\code{run_app} is a wrapper for the \code{epiworldRShiny} function. It is a
convenience function to run the app.
}
