% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-npi.R
\name{npi_add_masking}
\alias{npi_add_masking}
\title{Masking definition function}
\usage{
npi_add_masking(model, preval)
}
\arguments{
\item{model}{epiworldR model.}

\item{preval}{Prevalence of masking within the population.}
}
\value{
Creates a masking npi with specified prevalence in specified model.
}
\description{
Masking definition function
}
