% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-npi.R
\name{npi_add_school_closure}
\alias{npi_add_school_closure}
\title{School closure definition function}
\usage{
npi_add_school_closure(model, preval, day)
}
\arguments{
\item{model}{epiworldR model.}

\item{preval}{Prevalence of school closure within the population.}

\item{day}{Day in the simulation where school closure goes into effect.}
}
\value{
Creates a school closure npi with specified prevalence and day to
implement
}
\description{
School closure definition function
}
