% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-npi.R
\name{npi_add_vaccine}
\alias{npi_add_vaccine}
\title{Vaccine definition function}
\usage{
npi_add_vaccine(model, preval, sus_red, trans_red, rec_enhan, death_red)
}
\arguments{
\item{model}{epiworldR model.}

\item{preval}{Initial prevalence of the vaccine.}

\item{sus_red}{Reduction in susceptibility probability due
to vaccine.}

\item{trans_red}{Reduction in transmission probability due to
vaccine.}

\item{rec_enhan}{Probability increase in recovery due to vaccine.}

\item{death_red}{Reduction in death probability due to vaccine.}
}
\value{
Returns an object of class epiworld_model, where model is
substituted with the model name.
}
\description{
Vaccine definition function
}
\examples{
library(epiworldR) # for ModelSEIRCONN function
model <- ModelSEIRCONN("COVID-19", n = 1000, prevalence = 0.05,
                       contact_rate = 4, transmission_rate = 0.1,
                       incubation_days = 7, recovery_rate = 0.14)
run(model, ndays = 100, seed = 123)
npi_add_vaccine(model, preval = .8, susceptibility_reduction = .9,
                transmission_reduction = .5, recovery_enhancer = .5,
                death_reduction = .9)
}
\seealso{
Other npis: 
\code{\link{npi_add_all}()},
\code{\link{npi_add_masking}()},
\code{\link{npi_add_school_closure}()}
}
\concept{npis}
