% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{npis_input}
\alias{npis_input}
\title{npis_input Function}
\usage{
npis_input(model_name)
}
\arguments{
\item{model_name}{Name of the epiworldR model.}
}
\value{
Returns an object of class shiny.tag.list.
}
\description{
npis_input Function
}
\examples{
npis_input("SEIRD")
}
