% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{numeric_input_ndays}
\alias{numeric_input_ndays}
\title{numeric_input_ndays Function}
\usage{
numeric_input_ndays(model_name)
}
\value{
Allows for the ability to specify the number of simulation days
}
\description{
numeric_input_ndays Function
}
