% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-server.R
\name{plot_reproductive_epi}
\alias{plot_reproductive_epi}
\title{plot_reproductive_epi Function}
\usage{
plot_reproductive_epi(model)
}
\arguments{
\item{model}{The model object}
}
\value{
A plot displaying the reproductive number for the model over the
course of the simulation
}
\description{
This function generates a plot of the reproductive number over time
}
\examples{
library(epiworldR) # for ModelSEIR function
model <- ModelSEIR("COVID-19", prevalence = 0.7, transmission_rate = 0.3,
                   incubation_days = 6, recovery_rate = 0.15)
plot_reproductive_epi(model)
}
