% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{population_input}
\alias{population_input}
\title{population_input Function}
\usage{
population_input(model_name)
}
\value{
Generates population (equity) demographics
}
\description{
population_input Function
}
