% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{simulate_button}
\alias{simulate_button}
\title{simulate_button Function}
\usage{
simulate_button(model_name)
}
\value{
Runs the ABM simulation when selected
}
\description{
simulate_button Function
}
