% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{slider_input_rate}
\alias{slider_input_rate}
\title{slider_input_rate Function}
\usage{
slider_input_rate(model_name, rate_name, value, maxval = 1, input_label = NULL)
}
\value{
Creates all rate sliders
}
\description{
slider_input_rate Function
}
