% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{slider_prevalence}
\alias{slider_prevalence}
\title{slider_prevalence Function}
\usage{
slider_prevalence(model_name)
}
\value{
Creates the prevalence slider
}
\description{
slider_prevalence Function
}
