% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{text_input_disease_name}
\alias{text_input_disease_name}
\title{text_input_disease_name Function}
\usage{
text_input_disease_name(model_name)
}
\arguments{
\item{model_name}{Name of the epiworldR model.}
}
\value{
Creates the text input for the disease name
}
\description{
text_input_disease_name Function
}
