% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{epiworldRenv}
\alias{epiworldRenv}
\title{Access to the epiworldR environment.}
\usage{
epiworldRenv()
}
\value{
Returns the \code{epiworldR_env} environment.
}
\description{
This function is for internal use only.
}
