% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{epiworldRShiny}
\alias{epiworldRShiny}
\title{epiworldRShiny Function}
\usage{
epiworldRShiny(...)
}
\value{
Loads and opens the RShiny app for the epiworldR package
}
\description{
This function runs the Shiny app for epiworldR
}
