% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{epiworldRShiny}
\alias{epiworldRShiny}
\title{epiworldRShiny App Launcher}
\usage{
epiworldRShiny(...)
}
\arguments{
\item{...}{Currently ignored.}
}
\value{
Loads and opens the RShiny app for the epiworldR package
}
\description{
epiworldRShiny App Launcher
}
