% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{network_input}
\alias{network_input}
\title{network_input Function}
\usage{
network_input(model_name)
}
\arguments{
\item{model_name}{Name of the epiworldR model.}
}
\value{
Returns an object of class shiny.tag.list.
}
\description{
network_input Function
}
\examples{
network_input("SEIRD")
}
