% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-npi.R
\name{npi_add_all}
\alias{npi_add_all}
\title{NPI adding function}
\usage{
npi_add_all(model, modelname, input)
}
\arguments{
\item{model}{epiworldR model.}

\item{modelname}{Specified model.}

\item{input}{User epiworldR model selection.}
}
\value{
Returns an object of class epiworld_model, where model is
substituted with the model name.
}
\description{
NPI adding function
}
\keyword{internal}
