% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-npi.R
\name{npi_add_masking}
\alias{npi_add_masking}
\title{Masking definition function}
\usage{
npi_add_masking(model, preval)
}
\arguments{
\item{model}{epiworldR model.}

\item{preval}{Prevalence of masking within the population.}
}
\value{
Returns an object of class epiworld_model, where model is
substituted with the model name.
}
\description{
Masking definition function
}
\examples{
model <- ModelSEIR("COVID-19", prevalence = 0.7, transmission_rate = 0.3,
                   incubation_days = 6, recovery_rate = 0.15)
npi_add_masking(model, preval = .8)
}
