% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-npi.R
\name{npi_add_school_closure}
\alias{npi_add_school_closure}
\title{School closure definition function}
\usage{
npi_add_school_closure(model, preval, day)
}
\arguments{
\item{model}{epiworldR model.}

\item{preval}{Prevalence of school closure within the population.}

\item{day}{Day in the simulation where school closure goes into effect.}
}
\value{
Returns an object of class epiworld_model, where model is
substituted with the model name.
}
\description{
School closure definition function
}
\examples{
model <- ModelSEIR("COVID-19", prevalence = 0.7, transmission_rate = 0.3,
                   incubation_days = 6, recovery_rate = 0.15)
npi_add_school_closure(model, preval = .8, day = 10)
}
