% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-npi.R
\name{npi_add_vaccine}
\alias{npi_add_vaccine}
\title{Vaccine definition function}
\usage{
npi_add_vaccine(
  model,
  preval,
  susceptibility_reduction = 0.9,
  transmission_reduction = 0.5,
  recovery_enhancer = 0.5,
  death_reduction = 0.9
)
}
\arguments{
\item{model}{epiworldR model.}

\item{preval}{Initial prevalence of the vaccine.}

\item{susceptibility_reduction}{Reduction in susceptibility probability due
to vaccine.}

\item{transmission_reduction}{Reduction in transmission probability due to
vaccine.}

\item{recovery_enhancer}{Probability increase in recovery due to vaccine.}

\item{death_reduction}{Reduction in death probability due to vaccine.}
}
\value{
Returns a vaccine tool with parameter reducers/enhancers.
}
\description{
Vaccine definition function
}
