% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{numeric_input_ndays}
\alias{numeric_input_ndays}
\title{numeric_input_ndays Function}
\usage{
numeric_input_ndays(model_name)
}
\arguments{
\item{model_name}{Name of the epiworldR model.}
}
\value{
Returns an object of class shiny.tag.
}
\description{
numeric_input_ndays Function
}
\examples{
numeric_input_ndays("SEIRD")
}
