% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-server.R
\name{plot_epi}
\alias{plot_epi}
\title{plot_epi Function}
\usage{
plot_epi(model, mark_max)
}
\arguments{
\item{model}{The number of individuals in the population.}

\item{mark_max}{The state which will have a mark at the peak}
}
\value{
A plot displaying each state from the model over the course of the
simulation
}
\description{
This function generates a plot of the model states over time
}
\examples{
library(epiworldR) # for ModelSEIR function
model <- ModelSEIR("COVID-19", prevalence = 0.7, transmission_rate = 0.3,
                   incubation_days = 6, recovery_rate = 0.15)
plot_epi(model, mark_max = "Infected")

}
