% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{population_input}
\alias{population_input}
\title{population_input Function}
\usage{
population_input(model_name)
}
\arguments{
\item{model_name}{Name of the epiworldR model.}
}
\value{
Returns an object of class shiny.tag.list.
}
\description{
population_input Function
}
\examples{
population_input("SEIRD")
}
