% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{seed_input}
\alias{seed_input}
\title{seed_input Function}
\usage{
seed_input(model_name)
}
\arguments{
\item{model_name}{Name of the epiworldR model.}
}
\value{
Returns an object of class shiny.tag.
}
\description{
seed_input Function
}
\examples{
seed_input("SEIRD")
}
