% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{simulate_button}
\alias{simulate_button}
\title{simulate_button Function}
\usage{
simulate_button(model_name)
}
\arguments{
\item{model_name}{Name of the epiworldR model.}
}
\value{
Returns an object of class shiny.tag.
}
\description{
simulate_button Function
}
\examples{
simulate_button("SEIRD")
}
