% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{slider_input_rate}
\alias{slider_input_rate}
\title{slider_input_rate Function}
\usage{
slider_input_rate(model_name, rate_name, value, maxval = 1, input_label = NULL)
}
\arguments{
\item{model_name}{Name of the epiworldR model.}

\item{rate_name}{Name of the rate.}

\item{value}{Initial value for the slider.}

\item{maxval}{Maxiumum value for the slider.}

\item{input_label}{Aids in creating the appropriate slider name.}
}
\value{
Returns an object of class shiny.tag.
}
\description{
slider_input_rate Function
}
\examples{
slider_input_rate("SEIRD", "transmission", value = 0.3, maxval = 1,
input_label = NULL)
}
