% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ui.R
\name{slider_prevalence}
\alias{slider_prevalence}
\title{slider_prevalence Function}
\usage{
slider_prevalence(model_name)
}
\arguments{
\item{model_name}{Name of the epiworldR model.}
}
\value{
Returns an object of class shiny.tag.
}
\description{
slider_prevalence Function
}
\examples{
# slider_prevalence("SEIRD")
}
